+++ app/src/main/java/com/hongmyeoun/goldcalc/model/lostArkApi/APIRemote.kt
@@ -4,17 +4,17 @@ import android.content.Context
 import androidx.core.content.ContextCompat
 import com.google.gson.GsonBuilder
 import com.hongmyeoun.goldcalc.R
-import com.hongmyeoun.goldcalc.model.searchedInfo.card.CardDetail
-import com.hongmyeoun.goldcalc.model.searchedInfo.card.CardEffects
-import com.hongmyeoun.goldcalc.model.searchedInfo.card.Cards
-import com.hongmyeoun.goldcalc.model.searchedInfo.engravings.SkillEngravings
-import com.hongmyeoun.goldcalc.model.searchedInfo.engravings.SkillEngravingsDetail
-import com.hongmyeoun.goldcalc.model.searchedInfo.equipment.CharacterItem
-import com.hongmyeoun.goldcalc.model.searchedInfo.equipment.EquipmentDetail
-import com.hongmyeoun.goldcalc.model.searchedInfo.gem.Gem
-import com.hongmyeoun.goldcalc.model.searchedInfo.gem.GemDetail
-import com.hongmyeoun.goldcalc.model.searchedInfo.skills.CombatSkillsDetail
-import com.hongmyeoun.goldcalc.model.searchedInfo.skills.Skills
+import com.hongmyeoun.goldcalc.model.profile.card.CardDetail
+import com.hongmyeoun.goldcalc.model.profile.card.CardEffects
+import com.hongmyeoun.goldcalc.model.profile.card.Cards
+import com.hongmyeoun.goldcalc.model.profile.engravings.SkillEngravings
+import com.hongmyeoun.goldcalc.model.profile.engravings.SkillEngravingsDetail
+import com.hongmyeoun.goldcalc.model.profile.equipment.CharacterItem
+import com.hongmyeoun.goldcalc.model.profile.equipment.EquipmentDetail
+import com.hongmyeoun.goldcalc.model.profile.gem.Gem
+import com.hongmyeoun.goldcalc.model.profile.gem.GemDetail
+import com.hongmyeoun.goldcalc.model.profile.skills.CombatSkillsDetail
+import com.hongmyeoun.goldcalc.model.profile.skills.Skills
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import okhttp3.OkHttpClient
