+++ app/src/main/java/com/hongmyeoun/goldcalc/model/profile/engravings/SkillEngravingsDetail.kt
@@ -1,4 +1,4 @@
-package com.hongmyeoun.goldcalc.model.searchedInfo.engravings
+package com.hongmyeoun.goldcalc.model.profile.engravings
 
 import com.google.gson.JsonParser
