+++ app/src/main/java/com/hongmyeoun/goldcalc/model/profile/equipment/EquipmentDetail.kt
@@ -1,4 +1,4 @@
-package com.hongmyeoun.goldcalc.model.searchedInfo.equipment
+package com.hongmyeoun.goldcalc.model.profile.equipment
 
 import com.google.gson.JsonObject
 import com.google.gson.JsonParser
