+++ app/src/main/java/com/hongmyeoun/goldcalc/view/profile/content/card/Simple.kt
@@ -7,7 +7,7 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.hongmyeoun.goldcalc.model.searchedInfo.card.Cards
+import com.hongmyeoun.goldcalc.model.profile.card.Cards
 import com.hongmyeoun.goldcalc.viewModel.profile.CardVM
 
 @Composable
