+++ app/src/main/java/com/hongmyeoun/goldcalc/view/profile/content/equipments/detail/AccessoryDetail.kt
@@ -18,9 +18,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
-import com.hongmyeoun.goldcalc.model.searchedInfo.equipment.AbilityStone
-import com.hongmyeoun.goldcalc.model.searchedInfo.equipment.CharacterAccessory
-import com.hongmyeoun.goldcalc.model.searchedInfo.equipment.CharacterItem
+import com.hongmyeoun.goldcalc.model.profile.equipment.AbilityStone
+import com.hongmyeoun.goldcalc.model.profile.equipment.CharacterAccessory
+import com.hongmyeoun.goldcalc.model.profile.equipment.CharacterItem
 import com.hongmyeoun.goldcalc.ui.theme.ImageBG
 import com.hongmyeoun.goldcalc.ui.theme.LightGrayBG
 import com.hongmyeoun.goldcalc.ui.theme.RedQual
