+++ app/src/main/java/com/hongmyeoun/goldcalc/viewModel/homework/AbyssDungeonVM.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
-import com.hongmyeoun.goldcalc.model.goldCheck.AbyssDungeonModel
+import com.hongmyeoun.goldcalc.model.homework.AbyssDungeonModel
 import com.hongmyeoun.goldcalc.model.roomDB.character.Character
 
 class AbyssDungeonVM(val character: Character?): ViewModel() {
