+++ app/src/main/java/com/hongmyeoun/goldcalc/viewModel/homework/CommandBossVM.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
-import com.hongmyeoun.goldcalc.model.goldCheck.CommandBossModel
+import com.hongmyeoun.goldcalc.model.homework.CommandBossModel
 import com.hongmyeoun.goldcalc.model.roomDB.character.Character
 
 class CommandBossVM(val character: Character?): ViewModel() {
