+++ app/src/main/java/com/hongmyeoun/goldcalc/viewModel/homework/EpicRaidVM.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
-import com.hongmyeoun.goldcalc.model.goldCheck.EpicRaidModel
+import com.hongmyeoun.goldcalc.model.homework.EpicRaidModel
 import com.hongmyeoun.goldcalc.model.roomDB.character.Character
 
 class EpicRaidVM(val character: Character?): ViewModel() {
