+++ app/src/main/java/com/hongmyeoun/goldcalc/viewModel/homework/KazerothRaidVM.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
-import com.hongmyeoun.goldcalc.model.goldCheck.KazerothRaidModel
+import com.hongmyeoun.goldcalc.model.homework.KazerothRaidModel
 import com.hongmyeoun.goldcalc.model.roomDB.character.Character
 
 class KazerothRaidVM(val character: Character?) : ViewModel() {
