+++ app/src/main/java/com/hongmyeoun/goldcalc/viewModel/profile/ProfileVM.kt
@@ -12,12 +12,12 @@ import com.hongmyeoun.goldcalc.model.lostArkApi.APIRemote.getCharSkill
 import com.hongmyeoun.goldcalc.model.lostArkApi.CharacterDetail
 import com.hongmyeoun.goldcalc.model.roomDB.character.Character
 import com.hongmyeoun.goldcalc.model.roomDB.character.CharacterRepository
-import com.hongmyeoun.goldcalc.model.searchedInfo.card.CardEffects
-import com.hongmyeoun.goldcalc.model.searchedInfo.card.Cards
-import com.hongmyeoun.goldcalc.model.searchedInfo.engravings.SkillEngravings
-import com.hongmyeoun.goldcalc.model.searchedInfo.equipment.CharacterItem
-import com.hongmyeoun.goldcalc.model.searchedInfo.gem.Gem
-import com.hongmyeoun.goldcalc.model.searchedInfo.skills.Skills
+import com.hongmyeoun.goldcalc.model.profile.card.CardEffects
+import com.hongmyeoun.goldcalc.model.profile.card.Cards
+import com.hongmyeoun.goldcalc.model.profile.engravings.SkillEngravings
+import com.hongmyeoun.goldcalc.model.profile.equipment.CharacterItem
+import com.hongmyeoun.goldcalc.model.profile.gem.Gem
+import com.hongmyeoun.goldcalc.model.profile.skills.Skills
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
